Modify path label to display path on hovering
The label length caused viewport UI to unable to resize smaller past the length
of the path name of an image. The label will now trim text and show
ellipses instead if there is not enough space to display path label, and
hovering over the label shows the actual full path.